requirements.txt
langchain
wikipedia
langchain-community
langchain-openai-0.3.11
langchain-anthropic-0.3.10
python-dotenv-1.1.0
pydantic-2.11.1
duckduckgo-search
flask-3.1.0